requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
super-gradients>=3.0.0
Pillow>=8.0.0
numpy>=1.20.0
opencv-python>=4.5.0